requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
